make components and restructure
