+++ src/components/Stats.js
@@ -1,0 +1,524 @@
+import React from 'react'
+
+const Stats = () => {
+    return (
+        <div>
+           <div className="card">
+                  <div className="card-body">
+                    <div className="dropdown float-end">
+                      <a
+                        href="#"
+                        className="dropdown-toggle arrow-none card-drop"
+                        data-bs-toggle="dropdown"
+                        aria-expanded="false"
+                      >
+                        <i className="mdi mdi-dots-vertical"></i>
+                      </a>
+                      <div className="dropdown-menu dropdown-menu-end">
+                        <a href="#" className="dropdown-item">
+                          Action
+                        </a>
+
+                        <a href="#" className="dropdown-item">
+                          Another action
+                        </a>
+
+                        <a href="#" className="dropdown-item">
+                          Something else
+                        </a>
+
+                        <a href="#" className="dropdown-item">
+                          Separated link
+                        </a>
+                      </div>
+                    </div>
+                    <h4 className="header-title mt-0">Statistics</h4>
+                    <div
+                      id="morris-bar-example"
+                      dir="ltr"
+                      style={{
+                        height: "280px",
+                        position: "relative",
+                        WebkitTapHighlightColor: "rgba(0, 0, 0, 0)",
+                      }}
+                      className="morris-chart"
+                    >
+                      <svg
+                        height="280"
+                        version="1.1"
+                        width="705"
+                        xmlns="http://www.w3.org/2000/svg"
+                        xmlnsXlink="http://www.w3.org/1999/xlink"
+                        style={{
+                          overflow: "hidden",
+                          position: "relative",
+                          left: "-0.3875px",
+                          top: "-0.8px",
+                        }}
+                      >
+                        <desc
+                          style={{
+                            WebkitTapHighlightColor: "rgba(0, 0, 0, 0)",
+                          }}
+                        >
+                          Created with Raphaël 2.3.0
+                        </desc>
+                        <defs
+                          style={{
+                            WebkitTapHighlightColor: "rgba(0, 0, 0, 0)",
+                          }}
+                        ></defs>
+                        <text
+                          x="30.787988662719727"
+                          y="240.5999994277954"
+                          textAnchor="end"
+                          fontFamily="sans-serif"
+                          fontSize="12px"
+                          stroke="none"
+                          fill="#888888"
+                          style={{
+                            WebkitTapHighlightColor: "rgba(0, 0, 0, 0)",
+                            textAnchor: "end",
+                            fontFamily: "sans-serif",
+                            fontSize: "12px",
+                            fontWeight: "normal",
+                          }}
+                          fontWeight="normal"
+                        >
+                          <tspan
+                            style={{
+                              WebkitTapHighlightColor: "rgba(0, 0, 0, 0)",
+                            }}
+                            dy="3.9999899864196777"
+                          >
+                            0
+                          </tspan>
+                        </text>
+                        <path
+                          fill="none"
+                          stroke="#adb5bd"
+                          d="M43.28798866271973,240.5H680"
+                          strokeOpacity="0.1"
+                          strokeWidth="0.5"
+                          style={{
+                            WebkitTapHighlightColor: "rgba(0, 0, 0, 0)",
+                          }}
+                        ></path>
+                        <text
+                          x="30.787988662719727"
+                          y="186.69999957084656"
+                          textAnchor="end"
+                          fontFamily="sans-serif"
+                          fontSize="12px"
+                          stroke="none"
+                          fill="#888888"
+                          style={{
+                            WebkitTapHighlightColor: "rgba(0, 0, 0, 0)",
+                            textAnchor: "end",
+                            fontFamily: "sans-serif",
+                            fontSize: "12px",
+                            fontWeight: "normal",
+                          }}
+                          fontWeight="normal"
+                        >
+                          <tspan
+                            style={{
+                              WebkitTapHighlightColor: "rgba(0, 0, 0, 0)",
+                            }}
+                            dy="3.9999992847442627"
+                          >
+                            25
+                          </tspan>
+                        </text>
+                        <path
+                          fill="none"
+                          stroke="#adb5bd"
+                          d="M43.28798866271973,186.5H680"
+                          strokeOpacity="0.1"
+                          strokeWidth="0.5"
+                          style={{
+                            WebkitTapHighlightColor: "rgba(0, 0, 0, 0)",
+                          }}
+                        ></path>
+                        <text
+                          x="30.787988662719727"
+                          y="132.7999997138977"
+                          textAnchor="end"
+                          fontFamily="sans-serif"
+                          fontSize="12px"
+                          stroke="none"
+                          fill="#888888"
+                          style={{
+                            WebkitTapHighlightColor: "rgba(0, 0, 0, 0)",
+                            textAnchor: "end",
+                            fontFamily: "sans-serif",
+                            fontSize: "12px",
+                            fontWeight: "normal",
+                          }}
+                          fontWeight="normal"
+                        >
+                          <tspan
+                            style={{
+                              WebkitTapHighlightColor: "rgba(0, 0, 0, 0)",
+                            }}
+                            dy="4.000000953674316"
+                          >
+                            50
+                          </tspan>
+                        </text>
+                        <path
+                          fill="none"
+                          stroke="#adb5bd"
+                          d="M43.28798866271973,132.5H680"
+                          strokeOpacity="0.1"
+                          strokeWidth="0.5"
+                          style={{
+                            WebkitTapHighlightColor: "rgba(0, 0, 0, 0)",
+                          }}
+                        ></path>
+                        <text
+                          x="30.787988662719727"
+                          y="78.89999985694885"
+                          textAnchor="end"
+                          fontFamily="sans-serif"
+                          fontSize="12px"
+                          stroke="none"
+                          fill="#888888"
+                          style={{
+                            WebkitTapHighlightColor: "rgba(0, 0, 0, 0)",
+                            textAnchor: "end",
+                            fontFamily: "sans-serif",
+                            fontSize: "12px",
+                            fontWeight: "normal",
+                          }}
+                          fontWeight="normal"
+                        >
+                          <tspan
+                            style={{
+                              WebkitTapHighlightColor: "rgba(0, 0, 0, 0)",
+                            }}
+                            dy="3.999994993209839"
+                          >
+                            75
+                          </tspan>
+                        </text>
+                        <path
+                          fill="none"
+                          stroke="#adb5bd"
+                          d="M43.28798866271973,78.5H680"
+                          strokeOpacity="0.1"
+                          strokeWidth="0.5"
+                          style={{
+                            WebkitTapHighlightColor: "rgba(0, 0, 0, 0)",
+                          }}
+                        ></path>
+                        <text
+                          x="30.787988662719727"
+                          y="25"
+                          textAnchor="end"
+                          fontFamily="sans-serif"
+                          fontSize="12px"
+                          stroke="none"
+                          fill="#888888"
+                          style={{
+                            WebkitTapHighlightColor: "rgba(0, 0, 0, 0)",
+                            textAnchor: "end",
+                            fontFamily: "sans-serif",
+                            fontSize: "12px",
+                            fontWeight: "normal",
+                          }}
+                          fontWeight="normal"
+                        >
+                          <tspan
+                            style={{
+                              WebkitTapHighlightColor: "rgba(0, 0, 0, 0)",
+                            }}
+                            dy="3.999999523162842"
+                          >
+                            100
+                          </tspan>
+                        </text>
+                        <path
+                          fill="none"
+                          stroke="#adb5bd"
+                          d="M43.28798866271973,25.5H680"
+                          strokeOpacity="0.1"
+                          strokeWidth="0.5"
+                          style={{
+                            WebkitTapHighlightColor: "rgba(0, 0, 0, 0)",
+                          }}
+                        ></path>
+                        <text
+                          x="626.9406657218933"
+                          y="253.0999994277954"
+                          textAnchor="middle"
+                          fontFamily="sans-serif"
+                          fontSize="12px"
+                          stroke="none"
+                          fill="#888888"
+                          style={{
+                            WebkitTapHighlightColor: "rgba(0, 0, 0, 0)",
+                            textAnchor: "middle",
+                            fontFamily: "sans-serif",
+                            fontSize: "12px",
+                            fontWeight: "normal",
+                          }}
+                          fontWeight="normal"
+                          transform="matrix(1,0,0,1,0,7.2)"
+                        >
+                          <tspan
+                            style={{
+                              WebkitTapHighlightColor: "rgba(0, 0, 0, 0)",
+                            }}
+                            dy="3.9999899864196777"
+                          >
+                            2015
+                          </tspan>
+                        </text>
+                        <text
+                          x="520.8219971656799"
+                          y="253.0999994277954"
+                          textAnchor="middle"
+                          fontFamily="sans-serif"
+                          fontSize="12px"
+                          stroke="none"
+                          fill="#888888"
+                          style={{
+                            WebkitTapHighlightColor: "rgba(0, 0, 0, 0)",
+                            textAnchor: "middle",
+                            fontFamily: "sans-serif",
+                            fontSize: "12px",
+                            fontWeight: "normal",
+                          }}
+                          fontWeight="normal"
+                          transform="matrix(1,0,0,1,0,7.2)"
+                        >
+                          <tspan
+                            style={{
+                              WebkitTapHighlightColor: "rgba(0, 0, 0, 0)",
+                            }}
+                            dy="3.9999899864196777"
+                          >
+                            2014
+                          </tspan>
+                        </text>
+                        <text
+                          x="414.70332860946655"
+                          y="253.0999994277954"
+                          textAnchor="middle"
+                          fontFamily="sans-serif"
+                          fontSize="12px"
+                          stroke="none"
+                          fill="#888888"
+                          style={{
+                            WebkitTapHighlightColor: "rgba(0, 0, 0, 0)",
+                            textAnchor: "middle",
+                            fontFamily: "sans-serif",
+                            fontSize: "12px",
+                            fontWeight: "normal",
+                          }}
+                          fontWeight="normal"
+                          transform="matrix(1,0,0,1,0,7.2)"
+                        >
+                          <tspan
+                            style={{
+                              WebkitTapHighlightColor: "rgba(0, 0, 0, 0)",
+                            }}
+                            dy="3.9999899864196777"
+                          >
+                            2013
+                          </tspan>
+                        </text>
+                        <text
+                          x="308.5846600532532"
+                          y="253.0999994277954"
+                          textAnchor="middle"
+                          fontFamily="sans-serif"
+                          fontSize="12px"
+                          stroke="none"
+                          fill="#888888"
+                          style={{
+                            WebkitTapHighlightColor: "rgba(0, 0, 0, 0)",
+                            textAnchor: "middle",
+                            fontFamily: "sans-serif",
+                            fontSize: "12px",
+                            fontWeight: "normal",
+                          }}
+                          fontWeight="normal"
+                          transform="matrix(1,0,0,1,0,7.2)"
+                        >
+                          <tspan
+                            style={{
+                              WebkitTapHighlightColor: "rgba(0, 0, 0, 0)",
+                            }}
+                            dy="3.9999899864196777"
+                          >
+                            2012
+                          </tspan>
+                        </text>
+                        <text
+                          x="202.4659914970398"
+                          y="253.0999994277954"
+                          textAnchor="middle"
+                          fontFamily="sans-serif"
+                          fontSize="12px"
+                          stroke="none"
+                          fill="#888888"
+                          style={{
+                            WebkitTapHighlightColor: "rgba(0, 0, 0, 0)",
+                            textAnchor: "middle",
+                            fontFamily: "sans-serif",
+                            fontSize: "12px",
+                            fontWeight: "normal",
+                          }}
+                          fontWeight="normal"
+                          transform="matrix(1,0,0,1,0,7.2)"
+                        >
+                          <tspan
+                            style={{
+                              WebkitTapHighlightColor: "rgba(0, 0, 0, 0)",
+                            }}
+                            dy="3.9999899864196777"
+                          >
+                            2011
+                          </tspan>
+                        </text>
+                        <text
+                          x="96.34732294082642"
+                          y="253.0999994277954"
+                          textAnchor="middle"
+                          fontFamily="sans-serif"
+                          fontSize="12px"
+                          stroke="none"
+                          fill="#888888"
+                          style={{
+                            WebkitTapHighlightColor: "rgba(0, 0, 0, 0)",
+                            textAnchor: "middle",
+                            fontFamily: "sans-serif",
+                            fontSize: "12px",
+                            fontWeight: "normal",
+                          }}
+                          fontWeight="normal"
+                          transform="matrix(1,0,0,1,0,7.2)"
+                        >
+                          <tspan
+                            style={{
+                              WebkitTapHighlightColor: "rgba(0, 0, 0, 0)",
+                            }}
+                            dy="3.9999899864196777"
+                          >
+                            2010
+                          </tspan>
+                        </text>
+                        <rect
+                          x="85.73545608520507"
+                          y="78.89999985694885"
+                          width="21.223733711242676"
+                          height="161.69999957084656"
+                          rx="0"
+                          ry="0"
+                          fill="#188ae2"
+                          stroke="none"
+                          fillOpacity="1"
+                          style={{
+                            WebkitTapHighlightColor: "rgba(0, 0, 0, 0)",
+                            fillOpacity: "1",
+                          }}
+                        ></rect>
+                        <rect
+                          x="191.85412464141845"
+                          y="150.04799966812135"
+                          width="21.223733711242676"
+                          height="90.55199975967406"
+                          rx="0"
+                          ry="0"
+                          fill="#188ae2"
+                          stroke="none"
+                          fillOpacity="1"
+                          style={{
+                            WebkitTapHighlightColor: "rgba(0, 0, 0, 0)",
+                            fillOpacity: "1",
+                          }}
+                        ></rect>
+                        <rect
+                          x="297.97279319763186"
+                          y="78.89999985694885"
+                          width="21.223733711242676"
+                          height="161.69999957084656"
+                          rx="0"
+                          ry="0"
+                          fill="#188ae2"
+                          stroke="none"
+                          fillOpacity="1"
+                          style={{
+                            WebkitTapHighlightColor: "rgba(0, 0, 0, 0)",
+                            fillOpacity: "1",
+                          }}
+                        ></rect>
+                        <rect
+                          x="404.09146175384524"
+                          y="158.67199964523314"
+                          width="21.223733711242676"
+                          height="81.92799978256227"
+                          rx="0"
+                          ry="0"
+                          fill="#188ae2"
+                          stroke="none"
+                          fillOpacity="1"
+                          style={{
+                            WebkitTapHighlightColor: "rgba(0, 0, 0, 0)",
+                            fillOpacity: "1",
+                          }}
+                        ></rect>
+                        <rect
+                          x="510.2101303100586"
+                          y="199.63599953651428"
+                          width="21.223733711242676"
+                          height="40.963999891281134"
+                          rx="0"
+                          ry="0"
+                          fill="#188ae2"
+                          stroke="none"
+                          fillOpacity="1"
+                          style={{
+                            WebkitTapHighlightColor: "rgba(0, 0, 0, 0)",
+                            fillOpacity: "1",
+                          }}
+                        ></rect>
+                        <rect
+                          x="616.328798866272"
+                          y="40.09199995994567"
+                          width="21.223733711242676"
+                          height="200.50799946784974"
+                          rx="0"
+                          ry="0"
+                          fill="#188ae2"
+                          stroke="none"
+                          fillOpacity="1"
+                          style={{
+                            WebkitTapHighlightColor: "rgba(0, 0, 0, 0)",
+                            fillOpacity: "1",
+                          }}
+                        ></rect>
+                      </svg>
+                      <div
+                        className="morris-hover morris-default-style"
+                        style={{ left: "251px", top: "108px", display: "none" }}
+                      >
+                        <div className="morris-hover-row-label">2015</div>
+                        <div
+                          className="morris-hover-point"
+                          style={{ color: "#188ae2" }}
+                        >
+                          Statistics: 93
+                        </div>
+                      </div>
+                    </div>
+                  </div>
+                </div>
+            
+        </div>
+    )
+}
+
+export default Stats
